% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_helpers.R
\name{apde_caption}
\alias{apde_caption}
\title{Add a standard APDE caption to ggplot2 objects}
\usage{
apde_caption(data_source = "XYZ dataset", division = "Health Sciences, APDE")
}
\arguments{
\item{data_source}{Description of data source}

\item{division}{Division name}
}
\value{
A ggplot2 labs object
}
\description{
Add a standard APDE caption to ggplot2 objects
}
\examples{
\dontrun{
# Add caption to plot
ggplot(pressure, aes(temperature, pressure)) +
  geom_point() +
  apde_caption("Physical measurement data")
}

}
