% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{foreignbornDT}
\alias{foreignbornDT}
\title{Foreign Born Population by PUMA}
\format{
A data.table with foreign-born population estimates:
\describe{
\item{year}{Numeric. Year of the estimate (2022)}
\item{PUMACE20}{Numeric. PUMA identifier code}
\item{variable}{Character. Always "Foreign Born"}
\item{mean}{Numeric. Proportion of foreign-born residents}
}
}
\source{
2022 American Community Survey Public Use Microdata Sample (PUMS)
}
\usage{
data(foreignbornDT)
}
\description{
A dataset containing the proportion of foreign-born residents by Public Use
Microdata Area (PUMA) in King County, based on 2022 American Community Survey
PUMS data.
}
\details{
Dataset created using rads::calc() on ACS PUMS data.
Includes all King County PUMAs (codes 23301-23318).
}
\examples{
head(apde.graphs::foreignbornDT)

}
\keyword{datasets}
