% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_helpers.R
\name{rotate_x_labels}
\alias{rotate_x_labels}
\title{Rotate x-axis labels for better readability}
\usage{
rotate_x_labels(angle = 45, hjust = 1)
}
\arguments{
\item{angle}{Rotation angle in degrees (0-360)}

\item{hjust}{Horizontal justification (0-1)}
}
\value{
A ggplot2 theme object
}
\description{
Rotate x-axis labels for better readability
}
\examples{
\dontrun{
# Rotate x-axis labels
ggplot(pressure, aes(factor(temperature), pressure)) +
  geom_col() +
  rotate_x_labels()
}

}
