% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shapeBigCities}
\alias{shapeBigCities}
\title{King County Major Cities Shapefile}
\format{
An sf (simple features) object containing city boundaries:
\describe{
\item{NAME}{Character. Name of the city}
\item{geometry}{sfc_MULTIPOLYGON. Spatial geometry defining city boundaries}
}
}
\source{
Census 2020 shapefiles, accessed from APDE-CDIP/Shapefiles
}
\usage{
data(shapeBigCities)
}
\description{
A spatial dataset containing boundary polygons for major cities in King County.
The data is derived from 2020 Census shapefiles.
}
\details{
The dataset includes boundaries for eight major King County cities:
\itemize{
\item Auburn
\item Bellevue
\item Federal Way
\item Kent
\item Kirkland
\item Redmond
\item Renton
\item Seattle
}
}
\examples{
\dontrun{
# Basic plot of city boundaries
library(ggplot2)
library(sf)
ggplot(shapeBigCities) +
  geom_sf()
}

}
\keyword{datasets}
