% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shapeBigCities_centroids}
\alias{shapeBigCities_centroids}
\title{King County Major Cities Centroids}
\format{
An sf (simple features) object containing city centroids:
\describe{
\item{NAME}{Character. Name of the city}
\item{geometry}{sfc_POINT. Spatial geometry defining city centroids}
}
}
\source{
Derived from shapeBigCities using sf::st_centroid()
}
\usage{
data(shapeBigCities_centroids)
}
\description{
A spatial dataset containing centroid points for major cities in King County.
Derived from shapeBigCities.
}
\details{
Centroids are calculated from shapeBigCities using st_centroid().
Includes the same eight major King County cities.
}
\examples{
\dontrun{
# Plot city boundaries with centroid points
library(ggplot2)
library(sf)
ggplot() +
  geom_sf(data = shapeBigCities) +
  geom_sf(data = shapeBigCities_centroids, color = "red")
}

}
\keyword{datasets}
