% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shapeHRA20}
\alias{shapeHRA20}
\title{King County Health Reporting Areas (2020)}
\format{
An sf (simple features) object containing HRA boundaries:
\describe{
\item{HRA20}{Character. HRA name}
\item{geometry}{sfc_MULTIPOLYGON. Spatial geometry defining HRA boundaries}
}
}
\source{
APDE-CDIP/Shapefiles/HRA/hra_2020_nowater.shp
}
\usage{
data(shapeHRA20)
}
\description{
A spatial dataset containing boundary polygons for Health Reporting Areas (HRAs)
in King County, using 2020 vintage boundaries. Large water bodies are excluded.
}
\details{
Contains boundaries for all King County Health Reporting Areas,
excluding water bodies.
}
\examples{
\dontrun{
# Basic plot of HRA boundaries
library(ggplot2)
library(sf)
ggplot(shapeHRA20) +
  geom_sf()
}

}
\keyword{datasets}
