% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shapeInterstate}
\alias{shapeInterstate}
\title{Interstate Highways in King County}
\format{
An sf (simple features) object containing interstate highways:
\describe{
\item{Interstate}{Character. Highway name}
\item{geometry}{sfc_LINESTRING. Spatial geometry defining interstate highways}
}
}
\source{
tigris::primary_roads(year = 2022)
}
\usage{
data(shapeInterstate)
}
\description{
A spatial dataset containing linestrings for interstate highways crossing
through King County.
}
\details{
Contains I-5, I-90, and I-405
}
\examples{
\dontrun{
# Basic plot interstate highways
library(ggplot2)
library(sf)
ggplot(shapeInterstate) +
  geom_sf()
}

}
\keyword{datasets}
