% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shapeKC}
\alias{shapeKC}
\title{King County shapefile}
\format{
An sf (simple features) object containing interstate highways:
\describe{
\item{County}{Character. 'King'}
\item{geometry}{sfc_MULTIPOLYGON. Spatial geometry defining King County}
}
}
\source{
//dphcifs/APDE-CDIP/Shapefiles/Region/region_hra20_nowater.shp
}
\usage{
data(shapeKC)
}
\description{
A spatial dataset containing boundary polygons for King County. Large water
bodies are excluded.
}
\examples{
\dontrun{
# Basic plot of King County boundaries
library(ggplot2)
library(sf)
ggplot(shapeKC) +
  geom_sf()
}

}
\keyword{datasets}
