% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shapeMedicalCenters}
\alias{shapeMedicalCenters}
\title{King County Medical Centers}
\format{
An sf (simple features) object containing medical center points:
\describe{
\item{NAME}{Character. Name of the medical facility}
\item{geometry}{sfc_POINT. Spatial geometry defining medical center locations}
}
}
\source{
//Kcitfsrprpgdw01/kclib/Plibrary2/pubsafe/shapes/point/medical_facilities.shp
}
\usage{
data(shapeMedicalCenters)
}
\description{
A spatial dataset containing point locations for medical centers in King County.
}
\details{
Contains locations for medical facilities across King County.
}
\examples{
# Basic plot of medical center locations
library(ggplot2)
ggplot() +
  geom_sf(data = shapeKC) +
  geom_sf(data = shapeMedicalCenters, color = "red")

}
\keyword{datasets}
