% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shapeParcels}
\alias{shapeParcels}
\title{Sample of King County Residential Parcel coordinates}
\format{
An sf (simple features) object containing residential parcel points:
\describe{
\item{ParcelID}{Numeric. Parece; ID}
\item{geometry}{sfc_POINT. Spatial geometry defining the parcel}
}
}
\source{
kcparcelpop::pcoords
}
\usage{
data(shapeParcels)
}
\description{
A spatial dataset containing points for a 1/20th sample of King County
residential parcel point coordinates. Sampled on 1/6/2025.
}
\examples{
\dontrun{
# Basic plot of parcel coordinates
library(ggplot2)
library(sf)
ggplot() +
  geom_sf(data = shapeKC) +
  geom_sf(data = shapeParcels, color = "red")
}

}
\keyword{datasets}
