% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shapePUMA20}
\alias{shapePUMA20}
\title{King County PUMAs (2020)}
\format{
An sf (simple features) object containing PUMA boundaries:
\describe{
\item{PUMACE20}{Numeric. PUMA identifier code (23301-23318)}
\item{geometry}{sfc_MULTIPOLYGON. Spatial geometry defining PUMA boundaries}
}
}
\source{
Census 2020 shapefiles, accessed from APDE-CDIP/Shapefiles
}
\usage{
data(shapePUMA20)
}
\description{
A spatial dataset containing boundary polygons for Public Use Microdata Areas
(PUMAs) in King County, using 2020 Census definitions. Large water bodies are
excluded.
}
\details{
Contains boundaries for all King County PUMAs (codes 23301-23318).
}
\examples{
# Basic plot of PUMA boundaries
library(ggplot2)
ggplot(shapePUMA20) +
  geom_sf()

}
\keyword{datasets}
