% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shapeTraumaHospitals}
\alias{shapeTraumaHospitals}
\title{King County Acute Service (Trauma) Hospitals}
\format{
An sf (simple features) object containing hospital points:
\describe{
\item{NAME}{Character. Name of the hospital}
\item{geometry}{sfc_POINT. Spatial geometry defining hospital locations}
}
}
\source{
//Kcitfsrprpgdw01/kclib/Plibrary2/pubsafe/shapes/point/hospitals.shp
}
\usage{
data(shapeTraumaHospitals)
}
\description{
A spatial dataset containing point locations for trauma hospitals in King County.
}
\details{
Contains locations for acute service (trauma) hospitals in King County.
}
\examples{
\dontrun{
# Basic plot of trauma hospital locations
library(ggplot2)
library(sf)
ggplot() +
  geom_sf(data = shapeKC) +
  geom_sf(data = shapeTraumaHospitals, color = "red")
}

}
\keyword{datasets}
